use Objects.equals for better readability. use Byte.compare for better readability. use Character.compare for better readability. change from .longValue() to autoboxing for better readability.

+++ core/src/main/java/com/qq/tars/common/util/BeanAccessor.java
@@ -124,7 +124,7 @@ public class BeanAccessor {
         }
 
         Long result = offsetMap == null ? null : offsetMap.get(fieldName);
-        return result == null ? -1 : result.longValue();
+        return result == null ? -1 : result;
     }
 
     public static Object getBeanValue(Object bean, String fieldName) {
